refactor: Response 양식 수정, dto로 네이밍 변경
