+++ src/main/java/com/gdschongik/gdsc/domain/event/dto/EventParticipationDto.java
@@ -1,15 +1,13 @@
-package com.gdschongik.gdsc.domain.event.dto.response;
+package com.gdschongik.gdsc.domain.event.dto;
 
 import com.gdschongik.gdsc.domain.event.domain.AfterPartyApplicationStatus;
 import com.gdschongik.gdsc.domain.event.domain.AfterPartyAttendanceStatus;
 import com.gdschongik.gdsc.domain.event.domain.MainEventApplicationStatus;
 import com.gdschongik.gdsc.domain.event.domain.PaymentStatus;
 
-public record AfterPartyParticipationResponse(
+public record EventParticipationDto(
         Long eventParticipationId,
-        String participantName,
-        String participantStudentId,
-        String participantPhone,
+        ParticipantDto participant,
         Long memberId,
         MainEventApplicationStatus mainEventApplicationStatus,
         AfterPartyApplicationStatus afterPartyApplicationStatus,
