+++ src/main/java/com/gdschongik/gdsc/domain/event/dto/ParticipantDto.java
@@ -1,0 +1,3 @@
+package com.gdschongik.gdsc.domain.event.dto;
+
+public record ParticipantDto(String name, String studentId, String phone) {}
